untangle executor_linux.go
