+++ drivers/shared/executor/libcontainer_nsenter_linux.go
@@ -1,6 +1,8 @@
 // Copyright IBM Corp. 2015, 2025
 // SPDX-License-Identifier: MPL-2.0
 
+//go:build cgo
+
 package executor
 
 import (
